+++ swiyu-ts-verifier/src/test/java/ch/admin/bj/swiyu/tsverifier/statement/ProtectedVerificationAuthorizationTrustStatementTest.java
@@ -5,7 +5,6 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class ProtectedVerificationAuthorizationTrustStatementTest {
 
@@ -13,8 +12,10 @@ class ProtectedVerificationAuthorizationTrustStatementTest {
 
     @Test
     void testParseExample() throws JsonProcessingException {
-        var pvaTS = mapper.readValue(ExampleTrustStatement.pvaTS.getCombinedJson(), ProtectedVerificationAuthorizationTrustStatement.class);
-        assertThat(pvaTS.getTyp()).isEqualTo(StatementType.PROTECTED_VERIFICATION_AUTHORIZATION_TRUST_STATEMENT);
+        var pvaTS = mapper.readValue(ExampleTrustStatement.pvaTS.getBodyJson(), ProtectedVerificationAuthorizationTrustStatement.class);
+        var header = mapper.readValue(ExampleTrustStatement.pvaTS.getHeaderJson(), StatementHeader.class);
+        pvaTS.setStatementHeaders(header);
+        assertThat(pvaTS.getStatementHeaders().getTyp()).isEqualTo(StatementType.PROTECTED_VERIFICATION_AUTHORIZATION_TRUST_STATEMENT);
         assertThat(pvaTS.getAuthorizedFields()).hasSize(1).contains("personal_administrative_number");
     }
 }