+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [NEXT] – next
+- Separate header from body claims in trust-statement so that header claims and body claims can be distinguished and validated independently
+
 ## [1.6.2] – 2026-06-03
 - Bump didresolver to 2.8.0 for security updates.